+++ src/main/java/fr/traqueur/currencies/CurrencyProvider.java
@@ -2,12 +2,37 @@ package fr.traqueur.currencies;
 
 import org.bukkit.OfflinePlayer;
 
+import java.math.BigDecimal;
+
+/**
+ * Interface used to interact with a currency provider.
+ *
+ * @author Traqueur
+ */
 public interface CurrencyProvider {
 
-    void deposit(OfflinePlayer player, double amount);
+    /**
+     * Deposit a certain amount of currency to a player.
+     *
+     * @param player The player to deposit the money.
+     * @param amount The amount of currency to deposit.
+     */
+    void deposit(OfflinePlayer player, BigDecimal amount);
 
-    void withdraw(OfflinePlayer player, double amount);
+    /**
+     * Withdraw a certain amount of currency from a player.
+     *
+     * @param player The player to withdraw the money.
+     * @param amount The amount of currency to withdraw.
+     */
+    void withdraw(OfflinePlayer player, BigDecimal amount);
 
-    double getBalance(OfflinePlayer player);
+    /**
+     * Get the balance of a player.
+     *
+     * @param player The player to get the balance.
+     * @return The balance of the player.
+     */
+    BigDecimal getBalance(OfflinePlayer player);
 
 }
