+++ src/main/java/fr/traqueur/currencies/providers/ElementalGemsProvider.java
@@ -4,19 +4,21 @@ import fr.traqueur.currencies.CurrencyProvider;
 import me.elementalgaming.ElementalGems.GemAPI;
 import org.bukkit.OfflinePlayer;
 
+import java.math.BigDecimal;
+
 public class ElementalGemsProvider implements CurrencyProvider {
     @Override
-    public void deposit(OfflinePlayer player, double amount) {
-        GemAPI.addGems(player.getUniqueId(), (int) amount);
+    public void deposit(OfflinePlayer player, BigDecimal amount) {
+        GemAPI.addGems(player.getUniqueId(), amount.doubleValue());
     }
 
     @Override
-    public void withdraw(OfflinePlayer player, double amount) {
-        GemAPI.removeGems(player.getUniqueId(), (int) amount);
+    public void withdraw(OfflinePlayer player, BigDecimal amount) {
+        GemAPI.removeGems(player.getUniqueId(), amount.doubleValue());
     }
 
     @Override
-    public double getBalance(OfflinePlayer player) {
-        return GemAPI.getGems(player.getUniqueId());
+    public BigDecimal getBalance(OfflinePlayer player) {
+        return BigDecimal.valueOf(GemAPI.getGems(player.getUniqueId()));
     }
 }
