+++ src/main/java/fr/traqueur/currencies/providers/PlayerPointsProvider.java
@@ -6,12 +6,14 @@ import org.black_ixx.playerpoints.PlayerPointsAPI;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.plugin.java.JavaPlugin;
 
+import java.math.BigDecimal;
+
 public class PlayerPointsProvider implements CurrencyProvider {
 
     private PlayerPointsAPI playerPointsAPI;
 
     private PlayerPointsAPI getAPI() {
-        if(this.playerPointsAPI == null) {
+        if (this.playerPointsAPI == null) {
             PlayerPoints playerPoints = JavaPlugin.getPlugin(PlayerPoints.class);
             this.playerPointsAPI = playerPoints.getAPI();
         }
@@ -20,17 +22,17 @@ public class PlayerPointsProvider implements CurrencyProvider {
     }
 
     @Override
-    public void deposit(OfflinePlayer player, double amount) {
-        this.getAPI().give(player.getUniqueId(), (int) amount);
+    public void deposit(OfflinePlayer player, BigDecimal amount) {
+        this.getAPI().give(player.getUniqueId(), amount.intValue());
     }
 
     @Override
-    public void withdraw(OfflinePlayer player, double amount) {
-        this.getAPI().take(player.getUniqueId(), (int) amount);
+    public void withdraw(OfflinePlayer player, BigDecimal amount) {
+        this.getAPI().take(player.getUniqueId(), amount.intValue());
     }
 
     @Override
-    public double getBalance(OfflinePlayer player) {
-        return this.getAPI().look(player.getUniqueId());
+    public BigDecimal getBalance(OfflinePlayer player) {
+        return BigDecimal.valueOf(this.getAPI().look(player.getUniqueId()));
     }
 }
